+++ CHANGELOG.md
@@ -115,6 +115,7 @@
 - Update materialized_version for profile functions metrics ([#522](https://github.com/getsentry/vroom/pull/522))
 - Support writing functions metrics we extract from chunks into the functions dataset ([#524](https://github.com/getsentry/vroom/pull/524))
 - Keep top N samples in flamegraph. ([#526](https://github.com/getsentry/vroom/pull/526))
+- Add utility to merge a list of android chunks and generate a speedscope result ([#531](https://github.com/getsentry/vroom/pull/531))
 - Remove unused legacy flamegraph code path. ([#533](https://github.com/getsentry/vroom/pull/533))
 - Remove generic metrics ingestion ([#534](https://github.com/getsentry/vroom/pull/534))